pyproject.toml
[project]
name = "resybot"
version = "0.1.0"
description = "Resy reservation sniper bot"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1.0",
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = ["ruff", "mypy", "pytest"]

[project.scripts]
resybot = "resybot.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
